Heart-Disease-Prediction-System/requirements.txt
# Django and Web Framework
Django==5.0.1
djangorestframework==3.14.0

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Machine Learning (use versions compatible with App Engine)
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3

# Data Visualization (matplotlib-base is lighter)
matplotlib==3.8.2
seaborn==0.13.1

# Image Processing (for ECG analysis)
scikit-image==0.23.0
natsort==8.4.0

# Utilities
Pillow==10.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0

# For better performance
joblib==1.3.2

# Google Cloud specific
google-cloud-storage==2.14.0
